fixed the store
